refactor(#2287): move validation/ into parsers brick as sub-module
- Move src/nexus/validation/ → src/nexus/parsers/validation/ (git mv)
- Update all internal imports: nexus.validation.X → nexus.parsers.validation.X
- Replace validate_mount_path runtime cross-brick import with injected
  callback (constructor DI pattern per LEGO architecture)
- Update 5 test files with new import paths
- Update pyproject.toml mypy override for new module path
