+++ pyproject.toml
@@ -653,7 +653,7 @@ module = [
 disable_error_code = ["import-untyped", "no-any-return"]
 
 [[tool.mypy.overrides]]
-module = "nexus.validation.config"
+module = "nexus.parsers.validation.config"
 # PyYAML lazy import
 disable_error_code = ["import-untyped"]
