+++ src/nexus/parsers/validation/detector.py
