+++ src/nexus/parsers/validation/parsers/__init__.py
@@ -4,11 +4,11 @@ Each parser knows how to build a shell command and parse its output
 into structured ValidationError objects.
 """
 
-from nexus.validation.parsers.base import Validator
-from nexus.validation.parsers.clippy import CargoClippyValidator
-from nexus.validation.parsers.eslint import ESLintValidator
-from nexus.validation.parsers.mypy import MypyValidator
-from nexus.validation.parsers.ruff import RuffValidator
+from nexus.parsers.validation.parsers.base import Validator
+from nexus.parsers.validation.parsers.clippy import CargoClippyValidator
+from nexus.parsers.validation.parsers.eslint import ESLintValidator
+from nexus.parsers.validation.parsers.mypy import MypyValidator
+from nexus.parsers.validation.parsers.ruff import RuffValidator
 
 BUILTIN_VALIDATORS: dict[str, type[Validator]] = {
     "ruff": RuffValidator,
