+++ src/nexus/parsers/validation/parsers/mypy.py
@@ -10,8 +10,8 @@ import re
 import shlex
 from typing import Literal
 
-from nexus.validation.models import ValidationError, ValidatorConfig
-from nexus.validation.parsers.base import Validator
+from nexus.parsers.validation.models import ValidationError, ValidatorConfig
+from nexus.parsers.validation.parsers.base import Validator
 
 logger = logging.getLogger(__name__)
