+++ tests/e2e/self_contained/test_validation_pipeline.py
@@ -7,8 +7,8 @@ from unittest.mock import AsyncMock
 import pytest
 
 from nexus.bricks.sandbox.sandbox_provider import CodeExecutionResult
-from nexus.validation.models import ValidationPipelineConfig, ValidatorConfig
-from nexus.validation.runner import ValidationRunner
+from nexus.parsers.validation.models import ValidationPipelineConfig, ValidatorConfig
+from nexus.parsers.validation.runner import ValidationRunner
 
 
 def _make_provider(*responses: CodeExecutionResult) -> AsyncMock:
