+++ tests/e2e/self_contained/test_validation_rpc_dispatch.py
@@ -11,16 +11,16 @@ import dataclasses
 
 import pytest
 
-from nexus.server.protocol import METHOD_PARAMS, SandboxValidateParams
-from nexus.validation.models import (
+from nexus.parsers.validation.models import (
     ValidationError,
     ValidationResult,
     ValidatorConfig,
 )
-from nexus.validation.script_builder import (
+from nexus.parsers.validation.script_builder import (
     build_simple_validation_script,
     parse_simple_script_output,
 )
+from nexus.server.protocol import METHOD_PARAMS, SandboxValidateParams
 
 
 class TestSandboxValidateProtocol:
@@ -175,7 +175,7 @@ class TestPerformanceCharacteristics:
 
     def test_config_caching(self):
         """Config loader caches by key — second load is instant."""
-        from nexus.validation.config import ValidatorConfigLoader
+        from nexus.parsers.validation.config import ValidatorConfigLoader
 
         loader = ValidatorConfigLoader()
         yaml = "validators:\n  - name: ruff\n    command: ruff check .\n"
@@ -192,7 +192,7 @@ class TestPerformanceCharacteristics:
         from unittest.mock import AsyncMock
 
         from nexus.bricks.sandbox.sandbox_provider import CodeExecutionResult
-        from nexus.validation.detector import detect_project_validators
+        from nexus.parsers.validation.detector import detect_project_validators
 
         provider = AsyncMock()
         provider.run_code = AsyncMock(
