+++ tests/e2e/server/test_validation_e2e.py
@@ -19,9 +19,9 @@ import httpx
 import pytest
 
 from nexus.bricks.sandbox.sandbox_provider import CodeExecutionResult
-from nexus.validation.models import ValidationPipelineConfig, ValidatorConfig
-from nexus.validation.runner import ValidationRunner
-from nexus.validation.script_builder import (
+from nexus.parsers.validation.models import ValidationPipelineConfig, ValidatorConfig
+from nexus.parsers.validation.runner import ValidationRunner
+from nexus.parsers.validation.script_builder import (
     build_simple_validation_script,
     parse_simple_script_output,
 )
@@ -344,7 +344,7 @@ class TestPerformanceConstraints:
 
     def test_config_caching_avoids_reparse(self):
         """Config loader returns cached instance for same key."""
-        from nexus.validation.config import ValidatorConfigLoader
+        from nexus.parsers.validation.config import ValidatorConfigLoader
 
         yaml = (
             "validators:\n"
@@ -361,7 +361,7 @@ class TestPerformanceConstraints:
     @pytest.mark.asyncio
     async def test_detection_uses_single_ls_command(self):
         """Project detection is O(1) — single ls call, not per-file exists."""
-        from nexus.validation.detector import detect_project_validators
+        from nexus.parsers.validation.detector import detect_project_validators
 
         provider = AsyncMock()
         provider.run_code = AsyncMock(
