+++ tests/unit/parsers/test_validation_parsers.py
@@ -14,10 +14,10 @@ import json
 
 import pytest
 
-from nexus.validation.parsers.clippy import CargoClippyValidator
-from nexus.validation.parsers.eslint import ESLintValidator
-from nexus.validation.parsers.mypy import MypyValidator
-from nexus.validation.parsers.ruff import RuffValidator
+from nexus.parsers.validation.parsers.clippy import CargoClippyValidator
+from nexus.parsers.validation.parsers.eslint import ESLintValidator
+from nexus.parsers.validation.parsers.mypy import MypyValidator
+from nexus.parsers.validation.parsers.ruff import RuffValidator
 
 # ── Ruff Validator ────────────────────────────────────────────────
